data

+++ Project/Sources/simple_serial.h
@@ -42,7 +42,7 @@ struct MSG_BUTTONS{
 
 struct MSG_LIDAR{
   int sentinel;
-  unsigned long laserSample;
+  int laserSample;
   unsigned int last_sample_time;
   
 };
@@ -104,7 +104,7 @@ void SendGyroMsg(int rot_x, int rot_y, int rot_z);
 void SendButtonsMsg();
 void SendTextMsg(char* text_message); 
 
-void SendLidarMsg(unsigned long laser_reading); 
+void SendLidarMsg(int laser_reading); 
 void SendAngleMsg(int azimuth, int elevation);
